backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-cors
psycopg2-binary
python-dotenv
openpyxl
cryptography
fpdf2
boto3
